Sever created for Find-food app and loging page created
