+++ package.json
@@ -1,0 +1,23 @@
+{
+  "name": "find-food",
+  "version": "1.0.0",
+  "description": "restaurants\u001b[D\u001b[D\u001b[D\u001b[D\u001b[D\u001b[D\u001b[D\u001b[D\u001b[D\u001b[R\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[web application",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/HrithikKharvi/Find-food.git"
+  },
+  "author": "Hrithik, Arjun",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/HrithikKharvi/Find-food/issues"
+  },
+  "homepage": "https://github.com/HrithikKharvi/Find-food#readme",
+  "dependencies": {
+    "ejs": "^3.1.8",
+    "express": "^4.18.2"
+  }
+}
