+++ public/css/index.css
@@ -1,0 +1,110 @@
+*{
+    margin: 0;
+    padding: 0;
+    font-family: sans-serif;
+}
+.hero{
+    height: 100%;
+    width: 100%;
+    background-image: url('../images/banner2.jpg');
+    background-position: center;
+    background-size: cover;
+    position: absolute;
+
+}
+.form-box{
+    width:350px;
+    height: 480px;
+    position: relative;
+    margin: 6% auto;
+    background: rgba(245, 245, 245, 0.158);
+    padding: 5px;
+    overflow: hidden;
+    backface-visibility:hidden;
+    border-radius: 20px;
+}
+.button-box{
+    width:220px;
+    margin: 35px auto;
+    position: relative;
+    box-shadow: 0 0 20px 9px #ff61241f;
+    border-radius: 30px;
+}
+.toggle-btn{
+    padding: 10px 30px;
+    cursor: pointer;
+    background: transparent;
+    border: 0;
+    outline: none;
+    position: relative;
+}
+#btn{
+    top:0;
+    left:0;
+    position: absolute;
+    width: 110px;
+    height: 100%;
+    background: linear-gradient(to right,#ff105f,#ffad06);
+    border-radius: 30px;
+    transition: .5s;
+}
+.social-icons{
+    width: 30px auto;
+    text-align: center; 
+}
+.social-icons img{
+    width: 30px;
+    margin :0 12px;
+    box-shadow: 0 0 20px #7f7f7f3d;
+    cursor: pointer;
+    border-radius: 50%;
+}
+.input-group{
+    top:180px;
+    position:absolute;
+    width:280px;
+    transition: .5s;
+}
+.input-field{
+    width:100%;
+    padding: 10px 0;
+    margin :5px 0;
+    border-left: 0;
+    border-right: 0;
+    border-top: 0;
+    border-bottom: 1px solid #999;
+    outline: none;
+    background:transparent ;
+}
+.submit-btn{
+    width:85%;
+    padding:10px 30px;
+    display:block;
+    margin:auto;
+    background:linear-gradient(to right,#ff105f,#ffad06);
+    border:0;
+    outline: none;
+    border-radius: 30px;
+}
+.checkbox{
+    margin:30px 10px 30px 0;
+}
+span{
+    color:#777;
+    font-size:12px;
+    bottom: 68px;
+    position: absolute;
+}
+#login{
+    left:50px;
+
+}
+#register{
+    left:450px;
+    
+}
+p{
+    position:relative;
+    text-align: center;
+    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
+}
